move EndGame componentent into scenes/game
